live plotting option for all MontyExperiment instances
